Split JApplicationFwd.h from JApplication.h
